fix: 컴파일 오류 수정
- JobService.findByJobNo Long to Integer 변환
- Employment24ScraperController 미사용 import 제거
- PromptBuilder를 Job 모델 실제 필드에 맞게 수정

+++ src/main/java/com/seonier/service/JobService.java
@@ -21,7 +21,7 @@ public class JobService {
 
 
     public Job findByJobNo(Long jobNo) {
-        return jobMapper.findByJobNo(jobNo);
+        return jobMapper.findByJobNo(jobNo.intValue());
     }
 
     public List<Job> selectJobList() {
